requirements.txt
pygame-ce
sys